Get rid of "_logging" methods and traits

+++ src/apps/app/git_delta.rs
@@ -1,5 +1,10 @@
 use super::App;
-use crate::{install::IsInstalled, logging::HasLogger, Logger};
+use crate::{
+    actions::download,
+    install::{method::GithubRelease, CommandExists, Install},
+    os_package_managers::os_package_manager,
+    Logger, Success,
+};
 
 pub const ICON: char = '';
 
@@ -16,47 +21,50 @@ impl Default for GitDelta {
     }
 }
 
-impl HasLogger for GitDelta {
-    fn logger(&self) -> &Logger {
-        &self.logger
-    }
+impl CommandExists for GitDelta {
+    const CMD: &'static str = "delta";
 }
 
-impl IsInstalled for GitDelta {
-    fn is_installed(&self) -> bool {
-        crate::command_exists("delta")
-    }
-}
-
-#[cfg(target_os = "linux")]
-#[derive(Debug, thiserror::Error)]
-pub enum InstallFromGitHubReleaseError {
-    #[error("transparent")]
-    Download(#[from] crate::actions::download::Error),
-
-    #[error("transparent")]
-    IO(#[from] std::io::Error),
-}
-
-#[cfg(target_os = "linux")]
-impl crate::install::Install<crate::install::method::GitHubRelease<'_>> for GitDelta {
-    type Error = InstallFromGitHubReleaseError;
+impl Install<GithubRelease<'_>> for GitDelta {
+    type Error = Error;
 
     fn install(&self) -> Result<Success, Self::Error> {
-        use crate::os_package_managers::{os_package_manager::Dpkg, OsPackageManager};
+        self.logger
+            .log_sub_heading_group("install-via-github-release", || {
+        if cfg!(target_os = "linux") {
+            use crate::os_package_managers::{os_package_manager::Dpkg, OsPackageManager};
 
-        let deb_path = GitHubRelease::new(
-            "dandavison",
-            "delta",
-            "0.12.1",
-            "git-delta_0.12.1_amd64.deb",
-        )
-        .download()?;
-        Dpkg::default().install_package(&deb_path)?;
-        std::fs::remove_file(deb_path)?;
+            if Dpkg::command_exists() {
+                let deb_path = GithubRelease::new(
+                    "dandavison",
+                    "delta",
+                    "0.12.1",
+                    "git-delta_0.12.1_amd64.deb",
+                )
+                .download()?;
+
+                Dpkg::default().install_package(&deb_path)?;
+                std::fs::remove_file(deb_path)?;
+            } else {
+                todo!()
+            }
+        }
 
         Ok(Success::DidIt)
+            })
     }
 }
 
 impl App for GitDelta {}
+
+#[derive(Debug, thiserror::Error)]
+pub enum Error {
+    #[error("transparent")]
+    OsPackageManager(#[from] os_package_manager::Error),
+
+    #[error("transparent")]
+    Download(#[from] download::Error),
+
+    #[error("transparent")]
+    IO(#[from] std::io::Error),
+}
